LPS-35987 Optimize breadcrumb for Sites Administration
When navigating to see the list of subsites of a site, the nav-bar should also be visible. Since it already includes the breadcrumb it is not necessary to insert it manuallly. The case for the list of "root" sites will never show the breadcrumb anyway so nothing is lost with having the include inside (group != null)