LPS-37266 All ADD_PORTLET_DISPLAY_TEMPLATE permissions should be in its own resource
That way it's possible to show this permission only in the portlets that are meant to be in a page (as opposed to within the Control Panel or Site Administration) since those are the only ones where display templates is relevant. An additional benefit is that they can all be shown when configuring the permission of the "Application Display Templates" portlet itself.

Also shorten the name of the permission to "Add Display Template" since it's always shown in a context where the word Application is not necessary and could lead to confusion.