LPS-46832 Extract the processing of Control Panel redirects to its own method
This is the first step to invoke that method from a different place. This is necessary because the code inside the method added in the previous commits invokes several methods which require the ThemeDisplay in the request and it is not available until initThemeDisplay() finishes.
Since the goal is to move this method else where the goal has been to not pass any parameters other than request and response and make it more explicit that the method needs the ThemeDisplay available (ideally it would be explained in the Javadoc, but there are not javadocs in this class)