LPS-47465 Add a ServiceTrackerCustomizer to the ServiceTracker
This customizer will be the responsible to maintain the services in the
map.

The customizer returns the ServiceReference so the ServiceTracker will
track the references and provide the customizer with them. This way we
will use internally the references as handles and keep the services
ourselves to return them to the clients of the API.