LPS-47465 Don't use PropertyServiceReferenceComparator by default.
Instead use the inverse of the natural order of the service references.
The service references already are ordered using the _service.ranking_
property. We use the inverse so the first fetched elements are those
with highest ranking.

The behaviour of the map has not been changed. So the users may provide
their own comparator to the map. The PropertyServiceReferenceComparator
is left there since it can use any arbitrary property or logic to do the
ordering.