LPS-50156 move the entire package so we don't have a split package scenario
Note I deliberately kept the same package name for compatibility.
There's no reason to change it since in osgi we're exporting packages
of both portal-service and util-bridges together through the system bundle.
The difference however is that this package will no longer be duplicated
in traditional WAR plugins and so it will be possible to collaborate with
those plugins from osgi. Otherwise the duplicated class instance would
prevent this from working either by ClassCastException or complete ignorance
of published types.