LPS-52679 Add ExtensionManager to SoapExtender
We never remove an extension because it does not make sense from CXF
point of view, but the Apache Dependency Management API forces us to
provide add and remove methods. I have implemented an empty
removeExtension empty method in the ExtensionManager to signal this.

Sorry for the confusion this might have caused.