LPS-47563 Make DL search result tests pass
Change the implementation so that a summary is generated even if the
file entry doesn't exist.

In the expectations, allow multiple lookups of the AssetRendererFactory;
note that the summary will be generated only once anyway. The multiple
checking is a consequence of the factory returning null; once a correct
value is returned and a summary is generated, no more calls will occur.