LPS-56486 Fix blogs tests
Note that I had to force the call to `addComment` in
`PingbackMethodImpl` to return the value. Otherwise, Mockito would
incorrectly try to call a method with `void` return type, failing with a
`NoSuchMethodException` error.

This seems like a bug on Mockito. I'll investigate further; when (if) I
found a solution I'll remove this dummy return value.