LPS-59342 Default implementation
ConfigurationDescriptionFactoryImpl and
PropertiesFileNamedConfigurationContent are intentionally coupled
together so users can provide their own searching or configuration
interpretation strategies.

In the event that some users want/need to find configuration in a
different location or using a different language we need to make all of
them compatible because we don't want the original mechanism to stop
working.