LPS-63112 Exclude "buildThumbnails" on CI when only building CSS
This scenario happens in the "ruby-sass-compiler-jdk7" test batch. We can't call the Thumbnail Builder, because this test batch does not build the portal, so the portal artifacts are not available. It worked in the past because Gradle was able to find old portal artifact in the local or remote Maven repo.