LPS-66358 Use the correct asset renderer factory
Depending on the context, comments need to provide two different class
names: MBDiscussion and MBMessage. There is no easy solution to this, it
is an abstraction leak that we'll have to live with for a while.

In this particular case, the comments need to expose the MBMessage class
name for the workflow to work correctly, but we need to use the
MBDiscussion asset renderer factory to display it correctly. As the
asset API gets the asset renderer factory based on the asset class name,
we need to explicitly force it to use the Comment one (otherwise, the
base method will return the MBMessage one).