LPS-70464 Check class before cast
The CMIS Session is stored as a HTTP session attribute; if the module
is uninstalled and installed, the (CMIS) Session class will be
reloaded, and the implicit cast in `transientValue.getValue()` will
raise a ClassCastException.

Avoid the implicit cast by using a `Object` reference, and verify the
type before casting.